+++ .gitignore
@@ -1,0 +1,2 @@
+fenclaw_search
+saki_search
