--- config.py
@@ -1,17 +1,0 @@
-# File Configuration
-FILES = {
-    "search_log": "search_log.csv",
-    "usage_counter": "usage_counter.txt",
-    "master_list": "fenclaw_search/master_list.csv",
-    "with_emails": "fenclaw_search/stores_with_emails.csv",
-    "without_emails": "fenclaw_search/stores_without_email.csv",
-    "excluded_retailers": "fenclaw_search/excluded_retailers.csv"
-} 
-
-# FILES = {
-#     "search_log": "search_log.csv",
-#     "usage_counter": "usage_counter.txt",
-#     "master_list": "master_list.csv",
-#     "with_emails": "stores_with_emails.csv",
-#     "without_emails": "stores_without_email.csv"
-# }
