+++ shop_finder/config.py
@@ -1,0 +1,11 @@
+# File Configuration
+FILES = {
+    "search_log": "shop_finder/search_logs/search_log.csv",
+    "usage_counter": "shop_finder/search_logs/usage_counter.txt",
+    "master_list": "shop_finder/search_logs/master_list.csv",
+    "with_emails": "shop_finder/search_logs/stores_with_emails.csv",
+    "without_emails": "shop_finder/search_logs/stores_without_email.csv",
+    "excluded_retailers": "shop_finder/search_logs/excluded_retailers.csv",
+    "search_config" : "shop_finder/search_logs/search_config.csv",
+    "optimal_radii" : "shop_finder/search_logs/optimal_radii.csv"
+}
