+++ shop_finder/search_logs/Search_data_holder.csv
@@ -1,0 +1,1 @@
+Skip?,Status,city,lat,lng,radius,query
