+++ shop_finder/search_logs/excluded_retailers.csv
@@ -1,0 +1,1 @@
+retailer_name,reason
