+++ shop_finder/search_logs/optimal_radii.csv
