+++ shop_finder/search_logs/search_config.csv
@@ -1,0 +1,2 @@
+Skip?,Status,city,lat,lng,radius,query
+No,,Charlotte,35.2271,-80.8431,2,tabletop gaming store
