+++ shop_finder/search_logs/search_log.csv
@@ -1,0 +1,1 @@
+query,city,lat,lng,radius_m,date
