+++ shop_finder/search_logs/stores_with_emails.csv
@@ -1,0 +1,1 @@
+sorted,query,city,location_id,coordinates,name,address,website,email,flagged
