+++ shop_finder/search_logs/stores_without_email.csv
@@ -1,0 +1,1 @@
+sorted,query,city,location_id,coordinates,name,address,website,email,flagged
