+++ shop_finder/search_logs/usage_counter.txt
@@ -1,0 +1,1 @@
+0
