--- usage_counter.txt
@@ -1,1 +1,0 @@
-501
